Disabled quote correction for solhint test.

+++ test/fuzz/fuzz_helper/OverWriteFile.sol
@@ -103,7 +103,7 @@ contract OverWriteFile is PRBTest, StdCheats {
       return ""; // Return empty string if length is less than or equal to 2
     }
     bytes memory result = new bytes(strBytes.length - 2);
-    for (uint i = 1; i < strBytes.length - 1; i++) {
+    for (uint256 i = 1; i < strBytes.length - 1; i++) {
       result[i - 1] = strBytes[i];
     }
     return string(result);
